Fix support for escape and tab
When pressing tab, it should function like a select where
the currently selected item becomes the value and the select closes.
If no value is selected, the select will closes

Fix issue for escape not correctly closing the list.  It appeared
to be an issue with the callAll utility.  Replaced with calling each
function individually

Also fixes an issue where pressing the up/down arrow would not
re-open the list